dt-common: Macros for I2S pin labels on Nano board
For Nano (Porg) and Nano 2GB (Batuu) boards, macros are introduced to
set the pin label and pin group of I2S pins in Jetson-IO tool. This
allows a common DTSI file to be used for specifying the 40-pin header
overlay on all these boards.

Bug 200581790

Change-Id: I58de8366084f539af823d1e75e70f6597d12fb5b
Signed-off-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/tegra/+/2420844
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>